Check for (re)creation of properties
Still we haven't handled the situation that a property does not exist. The
DB makes no difference between a property that does not exist and a
property that has not been scanned for.

This could be solved with a second column for each property, but that
feels like a colossal waste. Maybe existence of the (undef) value in
the JSON would be a better indicator. If the path exists but the value
is undefined/undef , then the property does not exist for the file and
does not need to be rescanned...