A visit to the sentinel factory. We introduce our own EXISTS_BUT_EMPTY
This solves the problem of rescanning files while looking for tags
that have been checked but found to be empty.

"undef" / "NULL" can only represent one state, and we use that for
"was not checked for".

Maybe we should reverse the sense and use "\0unchecked" for all entries,
and then only update them with NULL / undef if they don't exist? But that
would make many queries ugly.